refactor(recipe/core): move RecipeRoot/Detail/Edit/AddGrocery screens to impl
RecipeRootScreen, RecipeDetailScreen, RecipeImageFullScreenScreen,
EditRecipeScreen, CategoryPickerSheet, CropPhotoOverlay, and
AddRecipeToGroceryListScreen all move into recipe/core/impl/
{root,detail,edit,addgrocery}/ui. RecipeDetailTestTags stays in
public for the recipe-detail robot.

RecipeRootBloc, RecipeDetailBloc, EditRecipeBloc, and
AddRecipeToGroceryListBloc extend BlocScreen; their Impls override
Content(). Inline preview fakes inside RecipeDetailScreen and
EditRecipeScreen gain Content() overrides.

RecipeRootBloc.Child sealed class itself extends BlocScreen so
RecipeRootScreen can collapse to `child.instance.Content()` — both
variants (Detail, Edit) provide it via `BlocScreen by bloc`.
RootBloc.Child.RecipeRoot delegates and RootScreen renders via
child.Content().

Snapshot tests updated: CategoryPickerScreenshotTest and
RecipeCategoriesRowScreenshotTest point at the new impl/ui
package; screenshot-test/build.gradle.kts adds recipe/core/impl.

recipe/core/public sets publicResClass = true; recipe/core/impl
gains ui/grocery-core/compose-resources alongside the already-
present Compose plugin.
