+++ client/recipe/core/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/impl/root/RecipeRootBlocImpl.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.recipe.core.impl.root
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.router.stack.StackNavigation
 import com.arkivanov.decompose.router.stack.bringToFront
@@ -13,6 +15,7 @@ import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.recipe.core.addmeal.MealPlannerRootBloc
 import com.plusmobileapps.chefmate.recipe.core.detail.RecipeDetailBloc
 import com.plusmobileapps.chefmate.recipe.core.edit.EditRecipeBloc
+import com.plusmobileapps.chefmate.recipe.core.impl.root.ui.RecipeRootScreen
 import com.plusmobileapps.chefmate.recipe.core.root.RecipeRootBloc
 import com.plusmobileapps.chefmate.recipe.data.ExtractedRecipeData
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
@@ -56,6 +59,11 @@ class RecipeRootBlocImpl(
 
     override fun onBackClicked() {
         navigation.pop()
+    }
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        RecipeRootScreen(recipeRootBloc = this, modifier = modifier)
     }
 
     private fun createChild(config: Configuration, context: BlocContext): RecipeRootBloc.Child =
