+++ client/recipe/core/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/impl/root/ui/RecipeRootScreen.kt
@@ -1,11 +1,10 @@
-package com.plusmobileapps.chefmate.recipe.core.root
+package com.plusmobileapps.chefmate.recipe.core.impl.root.ui
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.extensions.compose.stack.Children
-import com.plusmobileapps.chefmate.recipe.core.detail.RecipeDetailScreen
-import com.plusmobileapps.chefmate.recipe.core.edit.EditRecipeScreen
+import com.plusmobileapps.chefmate.recipe.core.root.RecipeRootBloc
 import com.plusmobileapps.chefmate.ui.backAnimation
 
 @Composable
@@ -19,9 +18,6 @@ fun RecipeRootScreen(recipeRootBloc: RecipeRootBloc, modifier: Modifier = Modifi
                 onBack = recipeRootBloc::onBackClicked,
             ),
     ) { child ->
-        when (val instance = child.instance) {
-            is RecipeRootBloc.Child.Detail -> RecipeDetailScreen(bloc = instance.bloc)
-            is RecipeRootBloc.Child.Edit -> EditRecipeScreen(bloc = instance.bloc)
-        }
+        child.instance.Content()
     }
 }
