+++ client/ui/screenshot-test/build.gradle.kts
@@ -43,6 +43,7 @@ dependencies {
     implementation(project(":client:grocery:core:public"))
     implementation(project(":client:grocery:core:impl"))
     implementation(project(":client:recipe:core:public"))
+    implementation(project(":client:recipe:core:impl"))
     implementation(project(":client:recipe:list:public"))
     implementation(project(":client:recipe:list:impl"))
     implementation(project(":client:recipe:data:public"))
